The main objective of this system, is to store and share our music to others. Using this system it makes our favorite music available everyw
This is a web based Music Library is the newest and most popular form of
listen music. In this system any user can listen and upload. The main
principle behind of this to share your music to others.
Here users can explore the music of his/her own and shared by others.
Users can create album, playlist and add their favorite music to it.
They can also share that album and playlist.